+++ server_tsspredator/server_visualization_processing.py
@@ -1,4 +1,4 @@
-from typing import Any, Literal
+from typing import  Literal
 import pandas as pd
 import os 
 import collections
